requirements_fixed.txt
# ====================================================================
# Sistema Gerencial - Backend Django + WhatsApp Integration
# Patrocínio, MG - Bruno
# Data: 10 de março de 2026
# ====================================================================

# ─────────────────────────────────────────────────────────────────
# Web Framework - Django + REST API
# ─────────────────────────────────────────────────────────────────
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.7
django-filter>=24

# ─────────────────────────────────────────────────────────────────
# Database - MySQL Connector
# ─────────────────────────────────────────────────────────────────
mysqlclient>=2.2

# ─────────────────────────────────────────────────────────────────
# Production Server & Static Files
# ─────────────────────────────────────────────────────────────────
gunicorn>=23.0
whitenoise>=6.8

# ─────────────────────────────────────────────────────────────────
# WhatsApp Automation via Playwright (WhatsApp Web)
# ─────────────────────────────────────────────────────────────────
playwright>=1.41.0
pytest-playwright>=0.4.0

# ─────────────────────────────────────────────────────────────────
# Utilities & Environment
# ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0
python-decouple>=3.8
requests>=2.31

# ─────────────────────────────────────────────────────────────────
# Image & QR Code Processing
# ─────────────────────────────────────────────────────────────────
Pillow>=11.0
qrcode[pil]>=8.0
python-barcode>=0.16

# ─────────────────────────────────────────────────────────────────
# Reports & Data Export
# ─────────────────────────────────────────────────────────────────
openpyxl>=3.1
reportlab>=4.2
pandas>=2.2.0

# ─────────────────────────────────────────────────────────────────
# XML & Digital Signature (NF-e, CT-e)
# ─────────────────────────────────────────────────────────────────
signxml>=3.1.0
lxml>=5.1.0

# ─────────────────────────────────────────────────────────────────
# Security & Cryptography
# ─────────────────────────────────────────────────────────────────
cryptography>=42.0.0

# ─────────────────────────────────────────────────────────────────
# AI & Machine Learning (Google Gen AI)
# ─────────────────────────────────────────────────────────────────
google-genai>=0.1.0
scikit-learn>=1.4.0

# ─────────────────────────────────────────────────────────────────
# Task Queue & Background Jobs (Recomendado para WhatsApp)
# ─────────────────────────────────────────────────────────────────
# celery>=5.3.0  # Descomente se quiser processar fila em background
# redis>=5.0.0  # Broker para Celery

# ─────────────────────────────────────────────────────────────────
# HTTP Utilities
# ─────────────────────────────────────────────────────────────────
httpx>=0.25.0
urllib3>=2.0.0
celery>=5.3.6
django-celery-beat>=2.5.0
redis>=5.0.1

# -----------------------------------------------------------------
# Task Queue & Async Processes
# -----------------------------------------------------------------
celery>=5.3.6
django-celery-beat>=2.5.0
redis>=5.0.1


